pyproject.toml
[tool.poetry]
name = "clipmath"
version = "0.1.0"
description = ""
authors = ["tpob <tpob@tpob.xyz>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
httpx = "^0.24.1"
pyperclip = "^1.8.2"
pillow = "^9.5.0"
pyyaml = "^6.0"
click = "^8.1.3"


[tool.poetry.group.dev.dependencies]
pyinstaller = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
